@@ -1,4 +1,12 @@
 import 'package:flutter/material.dart';
+import 'package:piwigo_ng/components/notification_dot.dart';
+import 'package:piwigo_ng/components/popup_list_item.dart';
+import 'package:piwigo_ng/services/preferences_service.dart';
+import 'package:piwigo_ng/services/upload_notifier.dart';
+import 'package:piwigo_ng/utils/localizations.dart';
+import 'package:piwigo_ng/views/image/image_favorites_page.dart';
+import 'package:piwigo_ng/views/upload/upload_status_page.dart';
+import 'package:provider/provider.dart';
 
 import '../../views/settings/settings_view_page.dart';
 import '../fields/app_field.dart';
@@ -32,9 +40,7 @@ class _RootSearchAppBarState extends State<RootSearchAppBar> {
       if (widget.scrollController.offset > _expandedHeight * _opacityScale) {
         return 0.0;
       }
-      return (_expandedHeight * _opacityScale -
-              widget.scrollController.offset) /
-          (_expandedHeight * _opacityScale);
+      return (_expandedHeight * _opacityScale - widget.scrollController.offset) / (_expandedHeight * _opacityScale);
     }
     return 1.0;
   }
@@ -50,8 +56,7 @@ class _RootSearchAppBarState extends State<RootSearchAppBar> {
       }
 
       // In case 0%-100% of the expanded height is viewed
-      double scrollDelta =
-          (_expandedHeight - widget.scrollController.offset) / _expandedHeight;
+      double scrollDelta = (_expandedHeight - widget.scrollController.offset) / _expandedHeight;
       double scrollPercent = (scrollDelta * 2 - 1);
       return (1 - scrollPercent) * delta * basePadding + basePadding;
     }
@@ -63,8 +68,7 @@ class _RootSearchAppBarState extends State<RootSearchAppBar> {
   Widget build(BuildContext context) {
     return SliverAppBar(
       leading: IconButton(
-        onPressed: () =>
-            Navigator.of(context).pushNamed(SettingsViewPage.routeName),
+        onPressed: () => Navigator.of(context).pushNamed(SettingsViewPage.routeName),
         icon: const Icon(Icons.settings),
       ),
       pinned: true,
@@ -75,12 +79,13 @@ class _RootSearchAppBarState extends State<RootSearchAppBar> {
         child: GestureDetector(
           onTap: widget.onSearch,
           child: const Hero(
-            tag: 'search-bar',
+            tag: '<search-bar>',
             child: Material(
               color: Colors.transparent,
               child: IgnorePointer(
                 child: AppField(
-                  icon: Icon(Icons.search),
+                  padding: const EdgeInsets.symmetric(vertical: 8.0),
+                  prefix: Icon(Icons.search),
                   hint: "Search...",
                 ),
               ),
@@ -89,10 +94,7 @@ class _RootSearchAppBarState extends State<RootSearchAppBar> {
         ),
       ),
       actions: [
-        IconButton(
-          onPressed: () {},
-          icon: const Icon(Icons.more_vert),
-        ),
+        _popupMenu,
       ],
       expandedHeight: _expandedHeight,
       flexibleSpace: FlexibleSpaceBar(
@@ -101,11 +103,67 @@ class _RootSearchAppBarState extends State<RootSearchAppBar> {
           vertical: 16,
         ),
         title: Text(
-          'Albums', // Todo: Use translations
+          appStrings.tabBar_albums,
           textScaleFactor: 1,
           style: Theme.of(context).appBarTheme.titleTextStyle,
         ),
       ),
+    );
+  }
+
+  Widget get _popupMenu {
+    return Stack(
+      alignment: Alignment.center,
+      children: [
+        PopupMenuButton(
+          position: PopupMenuPosition.under,
+          itemBuilder: (context) => [
+            PopupMenuItem(
+              onTap: () => Future.delayed(
+                const Duration(seconds: 0),
+                () => Navigator.of(context).pushNamed(UploadStatusPage.routeName),
+              ),
+              child: Stack(
+                children: [
+                  PopupListItem(
+                    icon: Icons.upload,
+                    text: appStrings.uploadSection_queue,
+                  ),
+                  Positioned(
+                    top: 14.0,
+                    left: 0.0,
+                    child: Consumer<UploadNotifier>(builder: (context, uploadNotifier, child) {
+                      return NotificationDot(
+                        isShown: uploadNotifier.uploadList.isNotEmpty,
+                      );
+                    }),
+                  ),
+                ],
+              ),
+            ),
+            if (Preferences.getUserStatus != 'guest')
+              PopupMenuItem(
+                onTap: () => Future.delayed(
+                  const Duration(seconds: 0),
+                  () => Navigator.of(context).pushNamed(ImageFavoritesPage.routeName),
+                ),
+                child: PopupListItem(
+                  icon: Icons.favorite,
+                  text: appStrings.categoryDiscoverFavorites_title,
+                ),
+              ),
+          ],
+        ),
+        Positioned(
+          top: 12.0,
+          left: 12.0,
+          child: Consumer<UploadNotifier>(builder: (context, uploadNotifier, child) {
+            return NotificationDot(
+              isShown: uploadNotifier.uploadList.isNotEmpty,
+            );
+          }),
+        ),
+      ],
     );
   }
 }
